+++ R/cps_download.R
@@ -2,11 +2,11 @@
 #' 
 #' @param path A file path (relative or absolute) where the downloads should go.
 #' @param years Which years of documentation to download. Defaults to all 
-#' even-numbered years from 1994 to 2022.
+#' even-numbered years from 1994 to 2024.
 #' @param overwrite Logical, whether to write over existing files or not. 
 #' Defaults to FALSE.
 #' @details 
-#' * File names will be written in the style "cps_nov2022.pdf", with the 
+#' * File names will be written in the style "cps_nov2024.pdf", with the 
 #' appropriate years.
 #' * The Voting and Registration Supplement is only conducted in even-numbered 
 #' years (since 1964), so any entry in `years` outside of this will be skipped.
@@ -19,7 +19,7 @@
 #' 
 #' @export
 cps_download_docs <- function(path = "cps_docs", 
-                          years = seq(1994, 2022, 2),
+                          years = seq(1994, 2024, 2),
                           overwrite = FALSE) {
   
   # sanitize inputs #####
@@ -39,24 +39,24 @@ cps_download_docs <- function(path = "cps_docs",
     years <- years[years >= 1994]
   }
   
-  # years must be before 2022
-  if (any(years > 2022)) {
-    warning(paste0("The Census Bureau has not yet released CPS data for years after 2022. The remaining years listed (",
-                   paste(years[years <= 2022], collapse = ", "),
+  # years must be before 2024
+  if (any(years > 2024)) {
+    warning(paste0("The Census Bureau has not yet released CPS data for years after 2024. The remaining years listed (",
+                   paste(years[years <= 2024], collapse = ", "),
                    ") will be downloaded."),
             immediate. = T)
-    years <- years[years <= 2022]
+    years <- years[years <= 2024]
   }
   
   # years must be in survey coverage zone
-  if (!all(years %in% seq(1964, 2022, 2))) {
+  if (!all(years %in% seq(1964, 2024, 2))) {
     warning(paste0("The VRS was not conducted in the following years: ",
-                   paste(years[!(years %in% seq(1964, 2022, 2))], collapse = ", "),
+                   paste(years[!(years %in% seq(1964, 2024, 2))], collapse = ", "),
                    ". The remaining years listed (",
-                   paste(years[years %in% seq(1964, 2022, 2)], collapse = ", "),
+                   paste(years[years %in% seq(1964, 2024, 2)], collapse = ", "),
                    ") will be downloaded."),
             immediate. = T)
-    years <- years[years %in% seq(1964, 2022, 2)]
+    years <- years[years %in% seq(1964, 2024, 2)]
   }
   
   # overwrite must be T/F
@@ -89,6 +89,7 @@ cps_download_docs <- function(path = "cps_docs",
   url_names <- dplyr::case_when(
     years < 2011 ~ paste0("data.nber.org/cps/cpsnov", stringr::str_sub(years, 3, 4), ".pdf"),
     years < 2017 ~ paste0("data.nber.org/cps/cpsnov", years, ".pdf"),
+    years == 2024 ~ paste0("https://www2.census.gov/programs-surveys/cps/techdocs/cpsnov", stringr::str_sub(years, 3, 4), ".pdf"),
     years > 2017 ~ paste0("data.nber.org/cps/cpsnov", stringr::str_sub(years, 3, 4), ".pdf")
   )
   
@@ -117,11 +118,11 @@ cps_download_docs <- function(path = "cps_docs",
 #' 
 #' @param path A file path (relative or absolute) where the downloads should go.
 #' @param years Which years of data to download. Defaults to all 
-#' even-numbered years from 1994 to 2022.
+#' even-numbered years from 1994 to 2024.
 #' @param overwrite Logical, whether to write over existing files or not. 
 #' Defaults to FALSE.
 #' @details 
-#' * File names will be written in the style "cps_nov2022.zip", with the 
+#' * File names will be written in the style "cps_nov2024.zip", with the 
 #' appropriate years.
 #' * The Voting and Registration Supplement is only conducted in even-numbered 
 #' years (since 1964), so any entry in `years` outside of this will be skipped.
@@ -134,7 +135,7 @@ cps_download_docs <- function(path = "cps_docs",
 #' 
 #' @export
 cps_download_data <- function(path = "cps_data", 
-                          years = seq(1994, 2022, 2), 
+                          years = seq(1994, 2024, 2), 
                           overwrite = FALSE) {
   
   # sanitize inputs #####
@@ -154,24 +155,24 @@ cps_download_data <- function(path = "cps_data",
     years <- years[years >= 1994]
   }
   
-  # years must be before 2022
-  if (any(years > 2022)) {
-    warning(paste0("The Census Bureau has not yet released CPS data for years after 2022. The remaining years listed (",
-                   paste(years[years <= 2022], collapse = ", "),
+  # years must be before 2024
+  if (any(years > 2024)) {
+    warning(paste0("The Census Bureau has not yet released CPS data for years after 2024. The remaining years listed (",
+                   paste(years[years <= 2024], collapse = ", "),
                    ") will be downloaded."),
             immediate. = T)
-    years <- years[years <= 2022]
+    years <- years[years <= 2024]
   }
   
   # years must be in survey coverage zone
-  if (!all(years %in% seq(1964, 2022, 2))) {
+  if (!all(years %in% seq(1964, 2024, 2))) {
     warning(paste0("The VRS was not conducted in the following years: ",
-                   paste(years[!(years %in% seq(1964, 2022, 2))], collapse = ", "),
+                   paste(years[!(years %in% seq(1964, 2024, 2))], collapse = ", "),
                    ". The remaining years listed (",
-                   paste(years[years %in% seq(1964, 2022, 2)], collapse = ", "),
+                   paste(years[years %in% seq(1964, 2024, 2)], collapse = ", "),
                    ") will be downloaded."),
             immediate. = T)
-    years <- years[years %in% seq(1964, 2022, 2)]
+    years <- years[years %in% seq(1964, 2024, 2)]
   }
   
   # overwrite must be T/F
@@ -205,6 +206,7 @@ cps_download_data <- function(path = "cps_data",
   url_names <- dplyr::case_when(
     years < 2011 ~ paste0("data.nber.org/cps/cpsnov", stringr::str_sub(years, 3, 4), ".zip"),
     years < 2017 ~ paste0("data.nber.org/cps/cpsnov", years, ".zip"),
+    years == 2024 ~ paste0("https://www2.census.gov/programs-surveys/cps/datasets/", years, "/supp/nov", stringr::str_sub(years, 3, 4), "pub.zip"),
     years > 2017 ~ paste0("data.nber.org/cps/nov", stringr::str_sub(years, 3, 4), "pub.zip")
   )